feat(admin): add empty edit page route for admin
This adds the initial empty page component for the admin edit route, which will be used to edit items by their ID. The route follows Next.js 13+ app directory conventions.
```

Note: Since the file is empty, I've included a body to explain the purpose of this addition. The commit type is "feat" as this is adding new functionality (a route) to the application. The scope is "admin" as this is clearly part of the admin functionality.